carrental/requirements.txt
Django==4.2.5
psycopg==3.1.10
psycopg2==2.9.9
pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-extra-views==0.14.0
django-allauth==0.56.1
django-filter==23.3
pytz==2023.3.post1
django-filter==23.3
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-datatables-too==0.1.4 
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
httpie==3.2.2
channels==3.0.5
channels-redis==4.1.0